==> src/f06_loopNum/step10_compareDist_all.R <==

# /home/toolkit/tools/R4.2.0/bin/R

library(Signac)
library(GenomicRanges)
library(Seurat)
#args <- commandArgs(TRUE)
#opath=args[1]
source('https://gitee.com/jumphone/BEER/raw/master/BEER.R')


.getYesNo<-function(DType,CType){
    DType=DType
    CType=CType
    i.yes=c()
    j.yes=c()
    i.no=c()
    j.no=c()
    i=1
    while(i<length(DType)){
        j=i+1
        while(j<=length(DType)){
            this_i_dtype=DType[i]
            this_j_dtype=DType[j]
            this_i_ctype=CType[i]
            this_j_ctype=CType[j]
            if(this_i_dtype != this_j_dtype & this_i_ctype == this_j_ctype){i.yes=c(i.yes, i);j.yes=c(j.yes, j)}
            if(this_i_ctype != this_j_ctype){i.no=c(i.no, i);j.no=c(j.no, j)}
            j=j+1}
        i=i+1}
    
    this_out=list()
    this_out$i.yes=i.yes
    this_out$j.yes=j.yes
    this_out$i.no=i.no
    this_out$j.no=j.no
    return(this_out)
    }


.getSub <-function(real.ref,pred.ref){
    COM=.simple_combine(real.ref,pred.ref)
    D1=COM$exp_sc_mat1
    D2=COM$exp_sc_mat2
    ALL=cbind(D1,D2)
    COR=cor(ALL,method='pearson')
    mat=COR
    DType=t(matrix(unlist(stringr::str_split(colnames(mat),'_')),nrow=2))[,2]
    CType=t(matrix(unlist(stringr::str_split(colnames(mat),'_')),nrow=2))[,1]
    this_out=.getYesNo(DType,CType)
    this_yes=diag(mat[this_out$i.yes, this_out$j.yes])
    this_no=diag(mat[this_out$i.no, this_out$j.no])
    #this_no_head=this_no[order(-this_no)][1:length(this_yes)]
    this_no_head=this_no[order(-this_no)][1]
    this_out=list()
    this_out$yes=this_yes
    this_out$no=this_no
    this_out$no_head=this_no_head
    #this_sub=mean(this_yes)-mean(this_no)
    return(this_out)
    }

#real.ref=readRDS('/home/disk/database/data/scHIC/GSE130711_analysis/NEW/COM_real.ref.rds')


OUT=list()
SUB=c()
################################################################
real.ref=readRDS('coolCount_100k/COM_real.ref.rds')
real.ref=real.ref[which(rowSums(real.ref)>0),]
pred.ref=readRDS('./ics_out/COM_pred.ref_100k.rds')
this_out=.getSub(real.ref,pred.ref)
OUT[[1]]=this_out$yes
OUT[[2]]=this_out$no_head
################################################################
real.ref=readRDS('coolCount_200k/COM_real.ref.rds')
real.ref=real.ref[which(rowSums(real.ref)>0),]
pred.ref=readRDS('./ics_out/COM_pred.ref_200k.rds')
this_out=.getSub(real.ref,pred.ref)
OUT[[3]]=this_out$yes
OUT[[4]]=this_out$no_head
################################################################
real.ref=readRDS('coolCount_300k/COM_real.ref.rds')
real.ref=real.ref[which(rowSums(real.ref)>0),]
pred.ref=readRDS('./ics_out/COM_pred.ref_300k.rds')
this_out=.getSub(real.ref,pred.ref)
OUT[[5]]=this_out$yes
OUT[[6]]=this_out$no_head
################################################################
real.ref=readRDS('coolCount_400k/COM_real.ref.rds')
real.ref=real.ref[which(rowSums(real.ref)>0),]
pred.ref=readRDS('./ics_out/COM_pred.ref_400k.rds')
this_out=.getSub(real.ref,pred.ref)
OUT[[7]]=this_out$yes
OUT[[8]]=this_out$no_head



SUB=c()
SUB=c(SUB,mean(OUT[[1]])-mean(OUT[[2]]))
SUB=c(SUB,mean(OUT[[3]])-mean(OUT[[4]]))
SUB=c(SUB,mean(OUT[[5]])-mean(OUT[[6]]))
SUB=c(SUB,mean(OUT[[7]])-mean(OUT[[8]]))

print(SUB)



pdf('f10p02_sub_barplot_cortex.pdf',width=2.5,height=5)

barplot(SUB,ylim=c(0,0.15),col=c(rep('grey60',0),rep('indianred1',4)),
        lwd=1,border = NA, space = 0.7)
abline(h=0,lwd=1.5)
dev.off()
